requirements.txt
numpy>=1.23,<3
PyYAML>=6.0
carla==0.9.16